+++ src/pedestrian/pedestrian-demo.cpp
@@ -5,11 +5,11 @@
 
 #include <opencv2/objdetect/objdetect.hpp>
 
-constexpr unsigned long WIDTH = 1920;
-constexpr unsigned long HEIGHT = 1080;
+constexpr unsigned int WIDTH = 1920;
+constexpr unsigned int HEIGHT = 1080;
 constexpr unsigned long DIAG = hypot(double(WIDTH), double(HEIGHT));
-constexpr unsigned long DOWNSIZE_WIDTH = 640;
-constexpr unsigned long DOWNSIZE_HEIGHT = 360;
+constexpr unsigned int DOWNSIZE_WIDTH = 640;
+constexpr unsigned int DOWNSIZE_HEIGHT = 360;
 constexpr double WIDTH_FACTOR = double(WIDTH) / DOWNSIZE_WIDTH;
 constexpr double HEIGHT_FACTOR = double(HEIGHT) / DOWNSIZE_HEIGHT;
 constexpr bool OFFSCREEN = false;
@@ -138,6 +138,7 @@ int main(int argc, char **argv) {
 
     if (!OFFSCREEN)
         x11::init("pedestrian-demo");
+    //you can set OpenGL-version, multisample-buffer samples and enable debug context using egl::init()
     egl::init();
     gl::init();
     nvg::init();
